+++ packages/@expo/config-plugins/src/ios/Updates.ts
@@ -1,6 +1,10 @@
+import { ExpoConfig } from '@expo/config-types';
+
+import { createBuildPodfilePropsConfigPlugin } from './BuildProperties';
 import { ExpoPlist } from './IosConfig.types';
 import { ConfigPlugin } from '../Plugin.types';
 import { withExpoPlist } from '../plugins/ios-plugins';
+import { withPlugins } from '../plugins/withPlugins';
 import {
   ExpoConfigUpdates,
   getDisableAntiBrickingMeasures,
@@ -34,6 +38,23 @@ export enum Config {
 // Also ensure the docs are up-to-date: https://docs.expo.dev/bare/installing-updates/
 
 export const withUpdates: ConfigPlugin = (config) => {
+  return withPlugins(config, [withUpdatesPlist, withUpdatesNativeDebugPodfileProps]);
+};
+
+/**
+ * A config-plugin to update `ios/Podfile.properties.json` from the `updates.useNativeDebug` in expo config
+ */
+export const withUpdatesNativeDebugPodfileProps = createBuildPodfilePropsConfigPlugin<ExpoConfig>(
+  [
+    {
+      propName: 'updatesNativeDebug',
+      propValueGetter: (config) => (config?.updates?.useNativeDebug === true ? 'true' : undefined),
+    },
+  ],
+  'withUpdatesNativeDebugPodfileProps'
+);
+
+const withUpdatesPlist: ConfigPlugin = (config) => {
   return withExpoPlist(config, async (config) => {
     const projectRoot = config.modRequest.projectRoot;
     const expoUpdatesPackageVersion = getExpoUpdatesPackageVersion(projectRoot);